Time strings support and readme
The timestrings were using the %l flag which
isn't supported by python in non unix environments
so I changed it to the %I flag which should be the
same value but cross platform. I added localhost to
the allowed host list, similar to the debug flag
in an ideal world this is removed before deploying
the actual site.

Also updated the readme with basic operation
instructions for developers new to Django.